split container
BREAKING CHANGE: split container to base ( containing resolve method ), event ( containing events and cqrs + base ) and light ( containing service + event )
