+++ src/types/StoreEntry.ts
@@ -1,0 +1,4 @@
+export interface StoreEntry {
+  type: "readable" | "writable" | "event" | "pipeline" | "handler" | "service";
+  value: unknown;
+}
